requirements.txt
google-cloud-firestore
playwright
bs4
requests